safari-600.7-branch/LayoutTests/media/media-source/media-source-canplaythrough-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
END OF TEST
